# edge-e2e/wrapper/src/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API definition for End-to-end testing of the Azure IoT SDKs.  All SDK APIs that are tested by our E2E tests need to be defined in this file.  This file takes some liberties with the API definitions.  In particular, response schemas are undefined, and error responses are also undefined.",
    "version" : "1.0.0",
    "title" : "IoT SDK Device & Client REST API"
  },
...
  "schemes" : [ "https" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/service/connect" : {
      "put" : {
        "tags" : [ "service" ],
        "summary" : "Connect to service",
        "description" : "Connect to the Azure IoTHub service.  More specifically, the SDK saves the connection string that is passed in for future use.",
        "parameters" : [ {
          "name" : "connectionString",
          "in" : "query",
          "description" : "Service connection string",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/service/{connectionId}/disconnect/" : {
      "put" : {
        "tags" : [ "service" ],
        "summary" : "Disconnect from the service",
        "description" : "Disconnects from the Azure IoTHub service.  More specifically, closes all connections and cleans up all resources for the active connection",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/service/{connectionId}/moduleMethod/{deviceId}/{moduleId}" : {
      "put" : {
        "tags" : [ "service" ],
        "summary" : "call the given method on the given module",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "moduleId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "methodInvokeParameters",
          "required" : true,
          "schema" : {
            "type" : "object"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object",
              "properties" : { }
            }
          }
        }
      }
    },
    "/service/{connectionId}/deviceMethod/{deviceId}" : {
      "put" : {
        "tags" : [ "service" ],
        "summary" : "call the given method on the given device",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "methodInvokeParameters",
          "required" : true,
          "schema" : {
            "type" : "object"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object",
              "properties" : { }
            }
          }
        }
      }
    },
...
    "/device/connect/{transportType}" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Connect to the azure IoT Hub as a device",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "transportType",
          "in" : "path",
          "description" : "Transport to use",
          "required" : true,
          "type" : "string",
          "enum" : [ "amqp", "amqpws", "mqtt", "mqttws", "http" ]
        }, {
          "name" : "connectionString",
          "in" : "query",
          "description" : "connection string",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "caCertificate",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/certificate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/device/{connectionId}/disconnect" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Disconnect the device",
        "description" : "Disconnects from Azure IoTHub service.  More specifically, closes all connections and cleans up all resources for the active connection",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/enableMethods" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Enable methods",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/methodRequest/{methodName}" : {
      "get" : {
        "tags" : [ "device" ],
        "summary" : "Wait for a method call with the given name",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "methodName",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/methodRequestResponse"
            }
          }
        }
      }
    },
    "/device/{connectionId}/methodResponse/{responseId}/{statusCode}" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Respond to the method call with the given name",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "responseId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "statusCode",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "responseBody",
          "required" : true,
          "schema" : {
            "type" : "object"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "definitions" : {
    "connectResponse" : {
      "type" : "object",
      "properties" : {
        "connectionId" : {
          "type" : "string",
          "description" : "unique identifier for this connection"
        }
      },
      "description" : "result of a connection to a service, device, or module client",
      "example" : {
        "connectionId" : "connectionId"
      }
    },
    "methodRequestResponse" : {
      "type" : "object",
      "properties" : {
        "requestPayload" : {
          "type" : "object",
          "description" : "payload for the request that arrived from the service",
          "properties" : { }
        },
        "responseId" : {
          "type" : "string",
          "description" : "response ID to be used when responding to this method call"
        }
      },
      "description" : "result of waiting on a method call",
      "example" : {
        "requestPayload" : "{}",
        "responseId" : "responseId"
      }
    },
...
    "certificate" : {
      "type" : "object",
      "properties" : {
        "cert" : {
          "type" : "string",
          "description" : "unique identifier for this connection"
        }
      },
      "description" : "certificate in the body of a message"
    }
  }
}